requirements.txt
streamlit-authenticator==0.2.3
snowflake-connector-python
pandas
streamlit
snowflake-snowpark-python
boto3
